Stripped the existing code and moved tests to the src directory
