+++ phpunit.xml
@@ -6,13 +6,16 @@
          bootstrap="vendor/autoload.php">
     <testsuites>
         <testsuite name="honest/bijective">
-            <directory suffix="Test.php">test</directory>
+            <directory suffix="Test.php">src/Tests</directory>
         </testsuite>
     </testsuites>
 
     <filter>
         <whitelist>
             <directory suffix=".php">src</directory>
+            <exclude>
+                <directory>src/Tests</directory>
+            </exclude>
         </whitelist>
     </filter>
 </phpunit>
