--- test/BijectiveTest.php
@@ -1,66 +1,0 @@
-<?php
-/**
- * Copyright © 2013 – 2014 Honest Empire Ltd
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of
- * this software and associated documentation files (the "Software"), to deal in the
- * Software without restriction, including without limitation the rights to use,
- * copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the
- * Software, and to permit persons to whom the Software is furnished to do so,
- * subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
- * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
- * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN
- * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
- * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- *
- * PHP version 5.3+
- *
- * @category  Bijection
- * @package   Bijective
- * @author    Daniel Morris <daniel@honestempire.com>
- * @copyright 2013 – 2014 Honest Empire Ltd
- * @license   MIT License
- * @link      https://github.com/honestempire/bijective
- */
-
-use Honest\Bijective\Bijective;
-
-/**
- * Test: Honest\Bijective\Bijective.php
- *
- * @category  Bijection
- * @package   Bijective
- * @author    Daniel Morris <daniel@honestempire.com>
- * @copyright 2013 – 2014 Honest Empire Ltd
- * @license   MIT License
- * @link      https://github.com/honestempire/bijective
- */
-class BijectiveTest extends PHPUnit_Framework_TestCase
-{
-
-    /**
-     * Honest\Bijective\Bijective::encode
-     * Honest\Bijective\Bijective::decode
-     *
-     * @covers Honest\Bijective\Bijective::encode
-     * @covers Honest\Bijective\Bijective::decode
-     * @return void
-     * @author Daniel Morris <daniel@honestempire.com>
-     * @access public
-     */
-    public function testEncodeDecode()
-    {
-        $max = mt_getrandmax();
-        for ($i = 0; $i < 1000; $i++) {
-            $rand = mt_rand(0, $max);
-            $this->assertSame($rand, Bijective::decode(Bijective::encode($rand)));
-        }
-    }
-
-}
